+++ R/connect.R
@@ -888,14 +888,31 @@ Connect <- R6::R6Class(
 
 #' Create a connection to Posit Connect
 #'
-#' Creates a connection to Posit Connect using the server URL and an api key.
+#' Creates a connection to Posit Connect using the server URL and an API key.
 #' Validates the connection and checks that the version of the server is
 #' compatible with the current version of the package.
+#'
+#' When running on Connect, the client's environment will contain default
+#' `CONNECT_SERVER` and `CONNECT_API_KEY` variables. The API key's permissions
+#' are scoped to the publishing user's account.
+#'
+#' To create a client with permissions scoped to the content visitor's account,
+#' call `connect()` passing a user session token from content session headers
+#' to the `token` argument. To do this, you must first add a Connect API
+#' integration in your published content's Access sidebar.
 #'
 #' @param server The URL for accessing Posit Connect. Defaults to environment
 #'   variable CONNECT_SERVER
 #' @param api_key The API Key to authenticate to Posit Connect with. Defaults
 #'   to environment variable CONNECT_API_KEY
+#' @param token Optional. A user session token. When running on a Connect server,
+#'   creates a client using the content visitor's account. Running locally, the
+#'   created client uses the provided API key.
+#' @param token_local_testing_key Optional. Only used when not running on
+#'   Connect and a `token` is provided. By default, the function returns a
+#'   `Connect` object using the `api_key`. By providing a different
+#'   key here you can test a visitor client with differently-scoped
+#'   permissions.
 #' @param prefix The prefix used to determine environment variables
 #' @param ... Additional arguments. Not used at present
 #' @param .check_is_fatal Whether to fail if "check" requests fail. Useful in
@@ -907,7 +924,17 @@ Connect <- R6::R6Class(
 #'
 #' @examples
 #' \dontrun{
-#' connect()
+#' client <- connect()
+#'
+#' # Running in Connect, create a client using the content visitor's account.
+#' # This example assumes code is being executed in a Shiny app's `server`
+#' # function with a `session` object available.
+#' token <- session$request$HTTP_POSIT_CONNECT_USER_SESSION_TOKEN
+#' client <- connect(token = token)
+#'
+#' # Test locally with an API key using a different role.
+#' fallback_key <- Sys.getenv("VIEWER_ROLE_API_KEY")
+#' client <- connect(token = token, token_local_testing_key = fallback_key)
 #' }
 #'
 #' @examplesIf identical(Sys.getenv("IN_PKGDOWN"), "true")
@@ -920,6 +947,8 @@ Connect <- R6::R6Class(
 connect <- function(
     server = Sys.getenv(paste0(prefix, "_SERVER"), NA_character_),
     api_key = Sys.getenv(paste0(prefix, "_API_KEY"), NA_character_),
+    token,
+    token_local_testing_key = api_key,
     prefix = "CONNECT",
     ...,
     .check_is_fatal = TRUE) {
@@ -932,6 +961,24 @@ connect <- function(
     }
   }
   con <- Connect$new(server = server, api_key = api_key)
+
+  if (!missing(token)) {
+    error_if_less_than(con$version, "2025.01.0")
+    if (on_connect()) {
+      visitor_creds <- get_oauth_credentials(
+        con,
+        user_session_token = token,
+        requested_token_type = "urn:posit:connect:api-key"
+      )
+      con <- connect(server = server, api_key = visitor_creds$access_token)
+    } else {
+      con <- connect(server = server, api_key = token_local_testing_key)
+      message(paste0(
+        "Called with `token` but not running on Connect. ",
+        "Continuing with fallback API key."
+      ))
+    }
+  }
 
   tryCatch(
     {