+++ tests/testthat/test-parse.R
@@ -25,7 +25,7 @@ test_that("coerce_datetime fills the void", {
   expect_error(coerce_datetime(NA_complex_, NA_datetime_, name = "complexity"), class = "vctrs_error_incompatible_type")
 })
 
-test_that("parse_connect_rfc3339 parses timestamps we expect from Connect", {
+test_that("parse_connect_rfc3339() parses timestamps with offsets as expected", {
   withr::defer(Sys.setenv(TZ = Sys.getenv("TZ")))
 
   x_mixed <- c(
@@ -82,6 +82,17 @@ test_that("parse_connect_rfc3339 parses timestamps we expect from Connect", {
   expect_identical(parse_connect_rfc3339(x_minus_one), expected)
   expect_identical(parse_connect_rfc3339(single_zero_offset), expected[1])
   expect_identical(parse_connect_rfc3339(single_offset), expected[1])
+})
+
+test_that("parse_connect_rfc3339() handles fractional seconds", {
+  expected <- as.POSIXct(strptime(c(
+    "2024-12-06T19:09:29.948016766+0000",
+    "2024-12-06T19:09:29.948070345+0000"
+  ), format = "%Y-%m-%dT%H:%M:%OS%z", tz = "UTC"))
+
+  x <- c("2024-12-06T19:09:29.948016766Z", "2024-12-06T19:09:29.948070345Z")
+
+  expect_identical(parse_connect_rfc3339(x), expected)
 })
 
 test_that("make_timestamp produces expected output", {
@@ -166,6 +177,7 @@ test_that("ensure_column works with lists", {
   expect_s3_class(list_chk_same, "tbl_df")
   expect_type(list_chk_same$hello, "list")
 })
+
 test_that("ensure_column works with POSIXct", {
   time_chk_null <- ensure_column(tibble::tibble(), NA_datetime_, "hello")
   expect_s3_class(time_chk_null, "tbl_df")